# spec/swagger.json
{"openapi":"3.0.0","paths":{"/version":{"get":{"operationId":"getVersion","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VersionEntity"}}}}},"tags":["app"]}},"/auth/access-token":{"get":{"operationId":"getAccessToken","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenEntity"}}}}},"tags":["auth"]}},"/auth/profile":{"get":{"operationId":"getProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileEntity"}}}}},"tags":["auth"]}},"/users/list":{"post":{"operationId":"listUsers","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiPaginatedRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseModel"}}}}},"tags":["users"]}},"/users/{userId}":{"get":{"operationId":"getUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"]},"put":{"operationId":"updateUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"]},"delete":{"operationId":"deleteUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["users"]}},"/users":{"post":{"operationId":"createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"]}},"/users/bulk":{"post":{"operationId":"createUsers","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BulkResponse"}}}}}},"tags":["users"]}},"/users/{userId}/copy":{"post":{"operationId":"copyUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"]}},"/images/{imageId}":{"get":{"operationId":"getImage","parameters":[{"name":"imageId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["images"]}},"/products/list":{"post":{"operationId":"listProducts","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiPaginatedRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponseModel"}}}}},"tags":["products"]}},"/products/featured":{"get":{"operationId":"getFeaturedProducts","parameters":[{"name":"visibleAmount","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductListEntity"}}}}}},"tags":["products"]}},"/products/{productId}":{"get":{"operationId":"getProduct","parameters":[{"name":"productId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductEntity"}}}}},"tags":["products"]}},"/products/3a9498d6-22ba-4ffa-ba00-3d68180d2618":{"post":{"operationId":"bulkImport","parameters":[],"responses":{"201":{"description":""}},"tags":["products"]}},"/categories":{"get":{"operationId":"getCategories","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponseModel"}}}}},"tags":["categories"]}},"/cart-items":{"get":{"operationId":"getCart","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartResponseModel"}}}}},"tags":["cartItems"]},"post":{"operationId":"add","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartItemAddEntity"}}}},"responses":{"204":{"description":""}},"tags":["cartItems"]}},"/cart-items/checkout":{"get":{"operationId":"getCartCheckout","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartCheckoutResponseModel"}}}}},"tags":["cartItems"]}},"/cart-items/{itemId}":{"delete":{"operationId":"remove","parameters":[{"name":"itemId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["cartItems"]},"put":{"operationId":"updateCount","parameters":[{"name":"itemId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartItemUpdateEntity"}}}},"responses":{"204":{"description":""}},"tags":["cartItems"]}}},"info":{"title":"API","description":"The main API for all Sandbox apps","version":"1.0","contact":{}},"tags":[{"name":"users","description":""},{"name":"accounts","description":""},{"name":"Differ","description":""},{"name":"Random","description":""}],"servers":[],"components":{"schemas":{"VersionEntity":{"type":"object","properties":{"version":{"type":"string"}},"required":["version"]},"AccessTokenEntity":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"ProfileEntity":{"type":"object","properties":{"sub":{"type":"string"},"family_name":{"type":"string"},"given_name":{"type":"string"},"email":{"type":"string"}},"required":["sub","family_name","given_name","email"]},"EnumFilterType":{"type":"string","enum":["Contains","GreaterThan","Search"]},"FilterType":{"type":"object","properties":{"value":{"oneOf":[{"type":"string"},{"type":"number"},{"type":"boolean"}]},"type":{"allOf":[{"$ref":"#/components/schemas/EnumFilterType"}]},"field":{"type":"string"}},"required":["value","type"]},"ApiPaginatedRequest":{"type":"object","properties":{"start":{"type":"number"},"limit":{"type":"number"},"page":{"type":"number"},"filters":{"type":"array","items":{"$ref":"#/components/schemas/FilterType"}}},"required":["start","limit","page"]},"AddressEntity":{"type":"object","properties":{"id":{"type":"number"},"lineOne":{"type":"string"},"lineTwo":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zipCode":{"type":"string"}},"required":["lineOne","city","state","zipCode"]},"UserEntity":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"birthDate":{"type":"number"},"gender":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressEntity"}},"required":["firstName","lastName","email"]},"UserResponseModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}},"total":{"type":"number"}},"required":["data"]},"BulkResponse":{"type":"object","properties":{"index":{"type":"number"},"message":{"type":"array","items":{"type":"string"}}},"required":["index","message"]},"CategoryEntity":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"parentId":{"type":"number"}},"required":["name"]},"ImageEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"contentType":{"type":"string"},"createdDate":{"type":"number"}},"required":["name","contentType","createdDate"]},"ProductListEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"price":{"type":"number"},"description":{"type":"string"},"category":{"$ref":"#/components/schemas/CategoryEntity"},"image":{"$ref":"#/components/schemas/ImageEntity"},"rating":{"type":"number"}},"required":["name","price","description","category","image","rating"]},"ProductResponseModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/ProductListEntity"}},"total":{"type":"number"}},"required":["data"]},"ReviewEntity":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"rating":{"type":"number"},"createdDate":{"type":"number"},"createdBy":{"$ref":"#/components/schemas/UserEntity"}},"required":["title","description","rating","createdDate","createdBy"]},"ProductEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"price":{"type":"number"},"description":{"type":"string"},"rating":{"type":"number"},"category":{"$ref":"#/components/schemas/CategoryEntity"},"image":{"$ref":"#/components/schemas/ImageEntity"},"reviews":{"type":"array","items":{"$ref":"#/components/schemas/ReviewEntity"}}},"required":["name","price","description","rating","category","image","reviews"]},"CategoryResponseModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CategoryEntity"}},"total":{"type":"number"}},"required":["data"]},"CartItemEntity":{"type":"object","properties":{"userId":{"type":"string"},"productId":{"type":"string"},"count":{"type":"number"},"createdDate":{"type":"number"},"product":{"$ref":"#/components/schemas/ProductListEntity"},"id":{"type":"number"}},"required":["userId","productId","count"]},"CartResponseModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CartItemEntity"}},"total":{"type":"number"}},"required":["data"]},"CartItemAddEntity":{"type":"object","properties":{"userId":{"type":"string"},"productId":{"type":"string"}},"required":["productId"]},"CartCheckoutItemEntity":{"type":"object","properties":{"userId":{"type":"string"},"productId":{"type":"string"},"count":{"type":"number"},"product":{"$ref":"#/components/schemas/ProductListEntity"},"subTotal":{"type":"number"},"id":{"type":"number"}},"required":["userId","productId","count","product","subTotal"]},"CartCheckoutResponseModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CartCheckoutItemEntity"}},"total":{"type":"number"},"subTotal":{"type":"number"},"tax":{"type":"number"},"taxPercent":{"type":"number"},"grandTotal":{"type":"number"},"shipping":{"type":"number"}},"required":["data","total","subTotal","tax","taxPercent","grandTotal","shipping"]},"CartItemUpdateEntity":{"type":"object","properties":{"total":{"type":"number"}},"required":["total"]}}}}